Fix eleven wire format bugs found by the full conformance factory sweep
Covering all 181 official OCPP v2.1 messages with serialization
factories surfaced another batch of copy-paste bugs, each contradicted
by sibling messages or the class's own parser:

- Seven identifiers/durations serialized as JSON strings or decimals
  where the official schemas demand integers: BatteryData.evseId,
  GetTariffsRequest.evseId, NotifyWebPaymentStartedRequest.timeout,
  ConstantStreamData.id and .variableMonitoringId,
  PullDynamicScheduleUpdateRequest.chargingProfileId,
  UpdateDynamicScheduleRequest.chargingProfileId and
  NotifyMonitoringReportRequest.requestId.

- All three tariff dimensions (TariffEnergy, TariffTime, TariffFixed)
  serialized and parsed "taxRate", but the official schemas define
  "taxRates" - the schema comment right next to the code already said
  so. First finding of the conformance suite's typo heuristic.

- SetVariableMonitoringResponse serialized its results under
  "setVariableResult" (copied from SetVariablesResponse), while its
  own parser already read the official "setMonitoringResult" - the
  roundtrip was provably broken.

- GetInstalledCertificateIdsResponse emitted an empty
  "certificateHashDataChain" array, violating minItems 1; it now
  follows the usual Any()-guard pattern.

Known findings (documented, not fixed - the data classes lack Edition 2
required properties): NotifyAllowedEnergyTransferRequest.transactionId,
VariableMonitoring.eventNotificationType and
NotifyPeriodicEventStream.pending/.basetime.

Tariff and signature fixtures of the main suite: 17/17 green.

+++ WWCP_OCPPv2.1/Messages/Charging/Transactions/NotifyWebPaymentStartedRequest.cs
@@ -376,7 +376,7 @@ namespace cloud.charging.open.protocols.OCPPv2_1.CSMS
                                : null,
 
                                  new JProperty("evseId",       EVSEId. Value),
-                                 new JProperty("timeout",      Timeout.TotalSeconds),
+                                 new JProperty("timeout",      (UInt64) Timeout.TotalSeconds),
 
                            Signatures.Any()
                                ? new JProperty("signatures",   new JArray(Signatures.Select(signature => signature.ToJSON(CustomSignatureSerializer,
